+++ recipes/essentials.rb
@@ -1,3 +1,5 @@
+node.override['build-essential']['compile_time'] = true
+
 include_recipe "yum"
 include_recipe "yum-epel"
 include_recipe "build-essential"