+++ recipes/gems.rb
@@ -1,0 +1,3 @@
+(node['cloudless-box']['gems'] || {}).map do |name, _|
+  gem_package name
+end