+++ recipes/packages.rb
@@ -1,0 +1,3 @@
+(node['cloudless-box']['packages'] || {}).map do |name, _|
+  package name
+end