+++ test/integration/default/serverspec/gems_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+describe 'gems' do
+  describe package('sass') do
+    it { should be_installed.by('gem') }
+  end
+end